+++ src/main/kotlin/no/nav/familie/pdf/pdf/språkKonfigurasjon/WebConfig.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.pdf.pdf.språkContext
+package no.nav.familie.pdf.pdf.språkKonfigurasjon
 
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry
